- (Core) Fixed issue where Page Properties block did not recognize context-aware Obsidian blocks.
