+++ Rock/Web/UI/RockBlock.cs
@@ -206,51 +206,7 @@ namespace Rock.Web.UI
         /// <summary>
         /// Gets a list of any context entities that the block requires.
         /// </summary>
-        public virtual List<EntityTypeCache> ContextTypesRequired
-        {
-            get
-            {
-                if ( _contextTypesRequired == null )
-                {
-                    _contextTypesRequired = new List<EntityTypeCache>();
-
-                    int properties = 0;
-                    foreach ( var attribute in this.GetType().GetCustomAttributes( typeof( ContextAwareAttribute ), true ) )
-                    {
-                        var contextAttribute = ( ContextAwareAttribute ) attribute;
-
-                        if ( !contextAttribute.Contexts.Any() )
-                        {
-                            // If the entity type was not specified in the attribute, look for a property that defines it
-                            string propertyKeyName = string.Format( "ContextEntityType{0}", properties > 0 ? properties.ToString() : string.Empty );
-                            properties++;
-
-                            Guid guid = Guid.Empty;
-                            if ( Guid.TryParse( GetAttributeValue( propertyKeyName ), out guid ) )
-                            {
-                                _contextTypesRequired.Add( EntityTypeCache.Get( guid ) );
-                            }
-                        }
-                        else
-                        {
-                            foreach ( var context in contextAttribute.Contexts )
-                            {
-                                var entityType = context.EntityType;
-
-                                if ( entityType != null && !_contextTypesRequired.Any( e => e.Guid.Equals( entityType.Guid ) ) )
-                                {
-                                    _contextTypesRequired.Add( entityType );
-                                }
-                            }
-                        }
-                    }
-                }
-
-                return _contextTypesRequired;
-            }
-        }
-
-        private List<EntityTypeCache> _contextTypesRequired;
+        public virtual List<EntityTypeCache> ContextTypesRequired => this.BlockCache?.ContextTypesRequired ?? new List<EntityTypeCache>();
 
         /// <summary>
         /// Gets a dictionary of the current context entities.  The key is the type of context, and the value is the entity object
