+++ RockWeb/Blocks/Mobile/MobilePageDetail.ascx.cs
@@ -813,40 +813,7 @@ namespace RockWeb.Blocks.Mobile
         /// </summary>
         private List<EntityTypeCache> GetContextTypesRequired( BlockCache block )
         {
-            var contextTypesRequired = new List<EntityTypeCache>();
-
-            int properties = 0;
-            foreach ( var attribute in block.BlockType.GetCompiledType().GetCustomAttributes( typeof( ContextAwareAttribute ), true ) )
-            {
-                var contextAttribute = ( ContextAwareAttribute ) attribute;
-
-                if ( !contextAttribute.Contexts.Any() )
-                {
-                    // If the entity type was not specified in the attribute, look for a property that defines it
-                    string propertyKeyName = string.Format( "ContextEntityType{0}", properties > 0 ? properties.ToString() : string.Empty );
-                    properties++;
-
-                    Guid guid = Guid.Empty;
-                    if ( Guid.TryParse( block.GetAttributeValue( propertyKeyName ), out guid ) )
-                    {
-                        contextTypesRequired.Add( EntityTypeCache.Get( guid ) );
-                    }
-                }
-                else
-                {
-                    foreach ( var context in contextAttribute.Contexts )
-                    {
-                        var entityType = context.EntityType;
-
-                        if ( entityType != null && !contextTypesRequired.Any( e => e.Guid.Equals( entityType.Guid ) ) )
-                        {
-                            contextTypesRequired.Add( entityType );
-                        }
-                    }
-                }
-            }
-
-            return contextTypesRequired;
+            return block?.ContextTypesRequired ?? new List<EntityTypeCache>();
         }
 
         /// <summary>
